Move SourceLevel under util package.
